+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.1
+// swift-tools-version:5.5
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -9,7 +9,6 @@ let package = Package(
         // Products define the executables and libraries produced by a package, and make them visible to other packages.
         .library(
             name: "ContentBlockerConverter",
-            type: .static,
             targets: ["ContentBlockerConverter", "FilterEngine"]),
         .executable(
             name: "ConverterTool",
@@ -24,20 +23,27 @@ let package = Package(
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages which this package depends on.
-        .target(
+        .executableTarget(
             name: "CommandLineWrapper",
-            dependencies: ["ContentBlockerConverter", "ArgumentParser"]),
+            dependencies: [
+                "ContentBlockerConverter",
+                .product(name: "ArgumentParser", package: "swift-argument-parser")
+            ]),
         .target(
             name: "ContentBlockerConverter",
-            dependencies: ["Punycode"]),
+            dependencies: [
+                .product(name: "Punycode", package: "PunycodeSwift")
+            ]),
         .target(
             name: "FilterEngine",
             dependencies: ["ContentBlockerConverter"]),
         .testTarget(
             name: "ContentBlockerConverterTests",
-            dependencies: ["ContentBlockerConverter"]),
+            dependencies: ["ContentBlockerConverter"],
+            resources: [.copy("Resources/test-rules.txt")]),
         .testTarget(
             name: "FilterEngineTests",
-            dependencies: ["FilterEngine"])
+            dependencies: ["FilterEngine"],
+            resources: [.copy("Resources/advanced-rules.txt")])
     ]
 )